@@ -11,7 +11,7 @@ from c8y_api.model import Application
 
 
 def test_parsing():
-    """Verify that parsing a Application from JSON works."""
+    """Verify that parsing an Application from JSON works."""
     path = os.path.dirname(__file__) + '/application.json'
     with open(path, encoding='utf-8', mode='rt') as f:
         application_json = json.load(f)
